Fix a possible warning on MSVC Win64

+++ src/opnmidi_opn2.cpp
@@ -315,7 +315,7 @@ void OPN2::Reset(int emulator, unsigned long PCM_RATE, void *audioTickHandler)
 #endif
         }
         cardsOP2[i].reset(chip);
-        chip->setChipId(i);
+        chip->setChipId((uint32_t)i);
         chip->setRate((uint32_t)PCM_RATE, 7670454);
         if(runAtPcmRate)
             chip->setRunningAtPcmRate(true);
